Fix integer overflow and support 64-bit integer numbers
- Upgrade GUI::SpinBox to inherit from QAbstractSpinBox instead of QSpinBox, unlocking native 64-bit support (qint64) for large integer inputs.
- Update NumberFieldWidget to initialize the spinbox with qint64 numeric limits and utilize toLongLong() instead of toInt() when parsing string values.
- Fix a fatal integer overflow crash in NumberComparison::compare by switching comparison floats to doubles and adding guards against int limits before calling qRound().

+++ src/gui/datewidget.cpp
@@ -105,7 +105,7 @@ DateWidget::DateWidget(QWidget* parent_) : QWidget(parent_) {
   l->addWidget(m_yearSpin, 1);
   l->setStretchFactor(m_yearSpin, 1);
 
-  void (SpinBox::* valueChangedInt)(int) = &SpinBox::valueChanged;
+  void (SpinBox::* valueChangedInt)(qint64) = &SpinBox::valueChanged;
   void (KComboBox::* indexChanged)(int) = &KComboBox::currentIndexChanged;
   connect(m_daySpin, valueChangedInt, this, &DateWidget::slotDateChanged);
   connect(m_monthCombo, indexChanged, this, &DateWidget::slotDateChanged);
